refactor : 렌쥬룰 판별 클래스 공통 로직 추상화 및 분리
