+++ src/test/kotlin/domain/board/rule/FourByFourTest.kt
@@ -1,0 +1,138 @@
+package domain.board.rule
+
+import domain.fixture.blackCByEight
+import domain.fixture.blackDByEight
+import domain.fixture.blackEByEight
+import domain.fixture.blackFByEight
+import domain.fixture.blackGByFive
+import domain.fixture.blackGByFour
+import domain.fixture.blackGBySeven
+import domain.fixture.blackGBySix
+import domain.fixture.emptyGByEight
+import domain.fixture.omokBoardFixture
+import domain.fixture.whiteCByEight
+import domain.fixture.whiteEByEight
+import domain.fixture.whiteFByEight
+import domain.fixture.whiteGByFour
+import domain.fixture.whiteGBySeven
+import domain.fixture.whiteGBySix
+import omok.domain.board.OmokBoard
+import omok.domain.rule.renju.FourByFour
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+
+class FourByFourTest {
+    private lateinit var board: OmokBoard
+
+    @BeforeEach
+    fun setUp() {
+        board = omokBoardFixture()
+    }
+
+    @Test
+    fun `4x4테스트1`() {
+        board.addStone(blackDByEight)
+        board.addStone(blackEByEight)
+        board.addStone(blackFByEight)
+        board.addStone(blackGBySeven)
+        board.addStone(blackGBySix)
+        board.addStone(blackGByFive)
+
+        val result = FourByFour(board).match(emptyGByEight)
+        assertThat(result).isTrue()
+    }
+
+    /**
+     *  15 ┌──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┐
+     *  14 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  13 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  12 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  11 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  10 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   9 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   8 ├──┼──●──●──●──┼──X──┼──┼──┼──┼──┼──┼──┼──┤
+     *   7 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   6 ├──┼──┼──┼──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┤
+     *   5 ├──┼──┼──┼──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┤
+     *   4 ├──┼──┼──┼──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┤
+     *   3 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   2 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   1 └──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┘
+     *     A  B  C  D  E  F  G  H  I  J  K  L  M  N  O
+     * */
+    @Test
+    fun `4x4테스트2`() {
+        board.addStone(blackCByEight)
+        board.addStone(blackDByEight)
+        board.addStone(blackEByEight)
+        board.addStone(blackGBySix)
+        board.addStone(blackGByFive)
+        board.addStone(blackGByFour)
+
+        val result = FourByFour(board).match(emptyGByEight)
+        assertThat(result).isTrue()
+    }
+
+    /**
+     *  15 ┌──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┐
+     *  14 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  13 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  12 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  11 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  10 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   9 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   8 ├──┼──●──┼──●──●──X──┼──┼──┼──┼──┼──┼──┼──┤
+     *   7 ├──┼──┼──┼──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┤
+     *   6 ├──┼──┼──┼──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┤
+     *   5 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   4 ├──┼──┼──┼──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┤
+     *   3 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   2 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   1 └──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┘
+     *     A  B  C  D  E  F  G  H  I  J  K  L  M  N  O
+     * */
+    @Test
+    fun `4x4테스트3`() {
+        board.addStone(blackCByEight)
+        board.addStone(blackEByEight)
+        board.addStone(blackFByEight)
+        board.addStone(blackGBySeven)
+        board.addStone(blackGBySix)
+        board.addStone(blackGByFour)
+
+        val result = FourByFour(board).match(emptyGByEight)
+        assertThat(result).isTrue()
+    }
+
+    /**
+     *  15 ┌──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┐
+     *  14 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  13 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  12 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  11 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  10 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   9 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   8 ├──┼──○──┼──○──○──X──┼──┼──┼──┼──┼──┼──┼──┤
+     *   7 ├──┼──┼──┼──┼──┼──○──┼──┼──┼──┼──┼──┼──┼──┤
+     *   6 ├──┼──┼──┼──┼──┼──○──┼──┼──┼──┼──┼──┼──┼──┤
+     *   5 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   4 ├──┼──┼──┼──┼──┼──○──┼──┼──┼──┼──┼──┼──┼──┤
+     *   3 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   2 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   1 └──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┘
+     *     A  B  C  D  E  F  G  H  I  J  K  L  M  N  O
+     * */
+    @Test
+    fun `4x4테스트4`() {
+        board.addStone(whiteCByEight)
+        board.addStone(whiteEByEight)
+        board.addStone(whiteFByEight)
+        board.addStone(whiteGBySeven)
+        board.addStone(whiteGBySix)
+        board.addStone(whiteGByFour)
+
+        val result = FourByFour(board).match(emptyGByEight)
+        assertThat(result).isFalse()
+    }
+}
