+++ src/test/kotlin/domain/board/rule/SixMokTest.kt
@@ -1,0 +1,88 @@
+package domain.board.rule
+
+import domain.fixture.blackBByEight
+import domain.fixture.blackCByEight
+import domain.fixture.blackDByEight
+import domain.fixture.blackEByEight
+import domain.fixture.blackGByEight
+import domain.fixture.emptyFByEight
+import domain.fixture.omokBoardFixture
+import domain.fixture.whiteBByEight
+import domain.fixture.whiteCByEight
+import domain.fixture.whiteDByEight
+import domain.fixture.whiteEByEight
+import domain.fixture.whiteGByEight
+import omok.domain.board.OmokBoard
+import omok.domain.rule.renju.SixMok
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+
+class SixMokTest {
+    private lateinit var board: OmokBoard
+
+    @BeforeEach
+    fun setUp() {
+        board = omokBoardFixture()
+    }
+
+    /**
+     *  15 ┌──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┐
+     *  14 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  13 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  12 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  11 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  10 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   9 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   8 ├──●──●──●──●──X──●──┼──┼──┼──┼──┼──┼──┼──┤
+     *   7 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   6 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   5 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   4 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   3 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   2 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   1 └──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┘
+     *     A  B  C  D  E  F  G  H  I  J  K  L  M  N  O
+     * */
+    @Test
+    fun `흑이 6목 이상의 장목이면 참을 반환한다`() {
+        board.addStone(blackBByEight)
+        board.addStone(blackCByEight)
+        board.addStone(blackDByEight)
+        board.addStone(blackEByEight)
+        board.addStone(blackGByEight)
+
+        val result = SixMok(board).match(emptyFByEight)
+        assertThat(result).isTrue()
+    }
+
+    /**
+     *  15 ┌──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┐
+     *  14 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  13 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  12 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  11 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *  10 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   9 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   8 ├──○──○──○──○──X──○──┼──┼──┼──┼──┼──┼──┼──┤
+     *   7 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   6 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   5 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   4 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   3 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   2 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
+     *   1 └──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┘
+     *     A  B  C  D  E  F  G  H  I  J  K  L  M  N  O
+     * */
+    @Test
+    fun `백이 6이상의 장목이면 거짓을 반환한다`() {
+        board.addStone(whiteBByEight)
+        board.addStone(whiteCByEight)
+        board.addStone(whiteDByEight)
+        board.addStone(whiteEByEight)
+        board.addStone(whiteGByEight)
+
+        val result = SixMok(board).match(emptyFByEight)
+        assertThat(result).isFalse()
+    }
+}
